+++ tests/Example/amphp-website/amp/coroutines/index.php
@@ -1,0 +1,10 @@
+<?php
+
+if (substr($_SERVER['REQUEST_URI'], -1, 1) !== '/') {
+    header('Location: ' . $_SERVER['REQUEST_URI'] . '/');
+    exit();
+}
+?>
+
+<html>
+</html>