+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## master
+
+- Upgrade from amphp/artax to amphp/http-client - @kelunik
+    - Adds `http_version` to report data
+- Fix end of link concat (#96) - @kelunik
+- Fix redirect URL resolution (#93) - @dantleech
+
 ## [0.9.0] 2019-06-08
 
 - Record referrer title if client encounters error while requesting URL, fixes #88